+++ gleam_stdlib/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `list` module gains `at`, `all`, `any`, `intersperse`, `unique`, `sort`, and
   `zip` functions.
+- `str` module gains `append`.
 
 ## v0.1.1 - 2019-04-17