+++ gleam_stdlib/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-- `list` module gains `at`, `all`, `any`, `intersperse`, `unique`, `sort`, and
-  `zip` functions.
+- `list` module gains `at`, `all`, `any`, `index_map`, `intersperse`, `sort`,
+  `unique`, and `zip` functions.
 - `str` module gains `append`.
 
 ## v0.1.1 - 2019-04-17